+++ Code/Views/ATLMessageCollectionViewCell.m
@@ -100,10 +100,11 @@ CGFloat const ATLMessageCellHorizontalMargin = 16.0f;
     // Remove self from any previously assigned LYRProgress instance.
     self.progress.delegate = nil;
     self.lastProgressFractionCompleted = 0;
+    [self.avatarImageView resetView];
     [self.bubbleView prepareForReuse];
 }
 
-- (void)presentMessage:(LYRMessage *)message;
+- (void)presentMessage:(LYRMessage *)message
 {
     self.message = message;
     LYRMessagePart *messagePart = message.parts.firstObject;